+++ RELEASENOTES.md
@@ -34,6 +34,8 @@
         `FLAG_READ_MFRA_FOR_SEEK_MAP` to the `FragmentedMp4Extractor`, which is
         now done by default in `DefaultExtractorsFactory`
         ([#3088](https://github.com/androidx/media/issues/3088)).
+    *   MP3: Ignore Xing data length if it's longer than the known stream length
+        ([#3117](https://github.com/androidx/media/issues/3117)).
 *   Inspector:
 *   Audio:
     *   Convert parameters of `AudioSink.configure` to data class. Custom